+++ changelog.md
@@ -8,6 +8,7 @@
 - Poly Fade
   - Removed audio pops when changing internal phasor direction
   - ***(breaking change)*** Ping pong internal phasor rate is now 1/2 displayed rate to preserve constant rate of channel switching
+  - New context menu option to disable "Reset if direction off"
 
 ### Bug Fixes
 - Repaired saved default themes system that was broken in v2.4.1